feat(api): Enhance API routes, response messages, and repository types

+++ back/src/main/java/com/openclassrooms/mddapi/repository/TopicRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TopicRepository extends JpaRepository<TopicEntity, Integer> {
+public interface TopicRepository extends JpaRepository<TopicEntity, Long> {
 }
